backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
websockets==12.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
Pillow==10.1.0
opencv-python==4.8.1.78
prophet==1.1.5
google-or-tools==9.8
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
bcrypt==4.0.1
qrcode[pil]==7.4.2
